requirements.txt
fastapi
openai
cohere
google-generativeai
anthropic
json
python-dotenv